Update slick.js (to further improve accessibility)
Hi I'm not accessibility expert but running the aXe https://chrome.google.com/webstore/detail/axe/lhdoppojpmngadmnindnejefpokejbdd it shows a "critical" error with regards to the attribute added to the button element/role "aria-selected". Looking at the ARIA spec I have changed it to "aria-pressed" as the best matched valid aria attribute for the element button (with a role of button).